refactor: pull AdHocSubProcessOutputCollectionBehavior into its own class
