+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/bpmn/container/AdHocSubProcessProcessor.java
@@ -11,6 +11,7 @@ import io.camunda.zeebe.el.Expression;
 import io.camunda.zeebe.engine.processing.adhocsubprocess.AdHocSubProcessUtils;
 import io.camunda.zeebe.engine.processing.bpmn.BpmnElementContainerProcessor;
 import io.camunda.zeebe.engine.processing.bpmn.BpmnElementContext;
+import io.camunda.zeebe.engine.processing.bpmn.behavior.AdHocSubProcessOutputCollectionBehavior;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnAdHocSubProcessBehavior;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnBehaviors;
 import io.camunda.zeebe.engine.processing.bpmn.behavior.BpmnCompensationSubscriptionBehaviour;
@@ -25,9 +26,6 @@ import io.camunda.zeebe.engine.processing.common.Failure;
 import io.camunda.zeebe.engine.processing.deployment.model.element.ExecutableAdHocSubProcess;
 import io.camunda.zeebe.model.bpmn.instance.zeebe.ZeebeAdHocImplementationType;
 import io.camunda.zeebe.msgpack.spec.MsgPackHelper;
-import io.camunda.zeebe.msgpack.spec.MsgPackReader;
-import io.camunda.zeebe.msgpack.spec.MsgPackType;
-import io.camunda.zeebe.msgpack.spec.MsgPackWriter;
 import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
 import io.camunda.zeebe.protocol.record.value.ErrorType;
 import io.camunda.zeebe.util.Either;
@@ -37,8 +35,6 @@ import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
 import org.agrona.DirectBuffer;
-import org.agrona.ExpandableArrayBuffer;
-import org.agrona.concurrent.UnsafeBuffer;
 
 public class AdHocSubProcessProcessor
     implements BpmnElementContainerProcessor<ExecutableAdHocSubProcess> {
@@ -313,44 +309,6 @@ public class AdHocSubProcessProcessor
             updatedCollection ->
                 stateBehavior.setLocalVariable(
                     adHocSubProcessContext, outputCollectionVariableName, updatedCollection));
-  }
-
-  private static final class AdHocSubProcessOutputCollectionBehavior {
-
-    private final MsgPackReader outputCollectionReader = new MsgPackReader();
-    private final MsgPackWriter outputCollectionWriter = new MsgPackWriter();
-    private final ExpandableArrayBuffer outputCollectionBuffer = new ExpandableArrayBuffer();
-    private final DirectBuffer updatedOutputCollectionBuffer = new UnsafeBuffer(0, 0);
-
-    public Either<Failure, DirectBuffer> appendToOutputCollection(
-        final DirectBuffer outputCollection, final DirectBuffer newValue) {
-
-      // read output collection
-      outputCollectionReader.wrap(outputCollection, 0, outputCollection.capacity());
-      final var token = outputCollectionReader.readToken();
-      if (token.getType() != MsgPackType.ARRAY) {
-        return Either.left(
-            new Failure(
-                "The output collection has the wrong type. Expected %s but was %s."
-                    .formatted(MsgPackType.ARRAY, token.getType())));
-      }
-      final int currentSize = token.getSize();
-      final int valuesOffset = outputCollectionReader.getOffset();
-
-      // write updated output collection
-      outputCollectionWriter.wrap(outputCollectionBuffer, 0);
-      outputCollectionWriter.writeArrayHeader(currentSize + 1);
-      // add current values
-      outputCollectionWriter.writeRaw(
-          outputCollection, valuesOffset, outputCollection.capacity() - valuesOffset);
-      // add new value
-      outputCollectionWriter.writeRaw(newValue);
-
-      final var length = outputCollectionWriter.getOffset();
-      updatedOutputCollectionBuffer.wrap(outputCollectionBuffer, 0, length);
-
-      return Either.right(updatedOutputCollectionBuffer);
-    }
   }
 
   private interface AdHocSubProcessBehavior {
